requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-devtools==1.0.post0
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
asyncmy==0.2.5
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
devtools==0.8.0
executing==0.8.3
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
iso8601==1.0.2
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
pycares==4.1.2
pycparser==2.21
Pygments==2.11.2
PyMySQL==1.0.2
pypika-tortoise==0.1.4
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
toml==0.10.2
tortoise-orm==0.19.0
typing_extensions==4.1.1
urllib3==1.26.9
watchgod==0.8.2
yarl==1.7.2
